add a hash function and switch to f64

+++ src/table.rs
@@ -7,11 +7,11 @@ struct LSHTable<'a, T: 'a, Q: 'a> {
     buckets: Vec<Bucket>,
     data: &'a Vec<T>,
     hash_functions: &'a Vec<&'a Q>,
-    ri1: Vec<f32>,
-    ri2: Vec<f32>
+    ri1: Vec<f64>,
+    ri2: Vec<f64>
 }
 
-impl<'a, T, Q: 'a> LSHTable<'a, T, Q> where Q: Fn(&'a T) -> f32 {
+impl<'a, T, Q: 'a> LSHTable<'a, T, Q> where Q: Fn(&'a T) -> f64 {
     pub fn new(data: &'a Vec<T>, hashes: &'a Vec<&'a Q>) -> Self {
         LSHTable {
             buckets: vec![Vec::new(); data.len() ],
@@ -24,7 +24,7 @@ impl<'a, T, Q: 'a> LSHTable<'a, T, Q> where Q: Fn(&'a T) -> f32 {
     pub fn new_build(data: &'a Vec<T>, hashes: &'a Vec<&'a Q>) -> Self {
         let mut x_to_build = LSHTable::new(data, hashes);
         for (i, v) in x_to_build.data.iter().enumerate() {
-            let hash_sig: Vec<f32> = hashes.iter().map(|x| {
+            let hash_sig: Vec<f64> = hashes.iter().map(|x| {
                 (*x)(v)
             }).collect();
             let bucket_ind = hash_func_t1(&hash_sig, &x_to_build.ri1, P, x_to_build.buckets.len());
@@ -34,7 +34,7 @@ impl<'a, T, Q: 'a> LSHTable<'a, T, Q> where Q: Fn(&'a T) -> f32 {
     }
 }
 
-fn hash_func_t1(signature: &Vec<f32>, rand_ints: &Vec<f32>, primes: i32, num_buckets: usize) -> usize {
+fn hash_func_t1(signature: &Vec<f64>, rand_ints: &Vec<f64>, primes: i32, num_buckets: usize) -> usize {
     let total: usize = signature.iter().zip(rand_ints).map(|(a, b)| {(a * b) as usize}).sum();
     (total % (P as usize)) % num_buckets
 }
@@ -50,7 +50,7 @@ mod tests {
             vec![1,2,3,4,5]
         ];
 
-        let val = |q: &Vec<i32>| {0.0 as f32};
+        let val = |q: &Vec<i32>| {0.0 as f64};
         let funcs = vec![&val];
         let x = LSHTable::new(&test_data, &funcs);
     }
@@ -62,7 +62,7 @@ mod tests {
             vec![1,2,3,4,5]
         ];
 
-        let val = |q: &Vec<i32>| {0.0 as f32};
+        let val = |q: &Vec<i32>| {0.0 as f64};
         let funcs = vec![&val];
         let x = LSHTable::new_build(&test_data, &funcs);
     }
